pyproject.toml
[project]
name = "selene"
version = "0.1.0"
description = "Selene AI - Death Dealer, Elder Slayer, Silent Watcher"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "leann>=0.3.7",
    "nuitka>=4.0.5",
    "ollama>=0.6.1",
    "orjson>=3.11.7",
    "pydantic-settings>=2.13.1",
    "structlog>=25.5.0",
    "tavily-python>=0.7.23",
    "textual>=8.1.1",
    "textual-serve>=1.1.3",
    "thoughtflow>=0.0.8",
    "typer-slim>=0.24.0",
]

[dependency-groups]
dev = [
    "prek>=0.3.6",
    "pytest>=9.0.2",
    "pytest-xdist>=3.8.0",
    "ruff>=0.15.6",
    "textual-dev>=1.8.0",
    "ty>=0.0.25",
]

[tool.pytest.ini_options]
testpaths = [ "src/tests" ]
pythonpath = [ "." ]

[tool.ruff]
lint.select = ["I"]
